Move approve + swap interaction to async thunk
This now runs entirely in the background script, ensuring the provider
isn't reinstantiated many times, but also allowing the UI to focus on
navigation.

Still missing: ensuring that after the approval transaction is signed
(if needed), the UI doesn't redirect to an asset page, but rather to the
next sign page.
